# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Rika Events - Enterprise Event Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues:
    get:
      operationId: getVenues
      tags: [venues]
      summary: List venues
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
    post:
      operationId: createVenue
      tags: [venues]
      summary: Create venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
  /venues/{id}:
    get:
      operationId: getVenue
      tags: [venues]
      summary: Get venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
    patch:
      operationId: updateVenue
      tags: [venues]
      summary: Update venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
    delete:
      operationId: deleteVenue
      tags: [venues]
      summary: Delete venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ROOMS ─────────────────────────────────────────────
...
  /vendors:
    get:
      operationId: getVendors
      tags: [vendors]
      summary: List vendors
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── GUESTS ────────────────────────────────────────────
...
components:
  schemas:
...
    Venue:
      type: object
      required: [id, name, location, capacity, pricePerDay, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        address:
          type: ["string", "null"]
        capacity:
          type: integer
        pricePerDay:
          type: number
        status:
          type: string
          enum: [available, booked, maintenance]
        amenities:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    VenueInput:
      type: object
      required: [name, location, capacity, pricePerDay]
      properties:
        name:
          type: string
        location:
          type: string
        address:
          type: string
        capacity:
          type: integer
        pricePerDay:
          type: number
        amenities:
          type: string
        imageUrl:
          type: string
        description:
          type: string
    VenueUpdate:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        address:
          type: string
        capacity:
          type: integer
        pricePerDay:
          type: number
        status:
          type: string
        amenities:
          type: string
        imageUrl:
          type: string
        description:
          type: string

    # ── ROOM ──────────────────────────────────────────
...
    Vendor:
      type: object
      required: [id, name, category, contactName, email, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [photography, catering, decor, music, transport, flowers, lighting, sound, video, makeup, security, other]
        contactName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, blacklisted]
        rating:
          type: ["number", "null"]
        totalEvents:
          type: ["integer", "null"]
        basePrice:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    VendorInput:
      type: object
      required: [name, category, contactName, email]
      properties:
        name:
          type: string
        category:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        basePrice:
          type: number
        notes:
          type: string
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        rating:
          type: number
        basePrice:
          type: number
        notes:
          type: string

    # ── GUEST ─────────────────────────────────────────
...