# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Rika Events - Enterprise Event Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VenueStatus } from './venueStatus';

export interface Venue {
  id: number;
  name: string;
  location: string;
  /** @nullable */
  address?: string | null;
  capacity: number;
  pricePerDay: number;
  status: VenueStatus;
  /** @nullable */
  amenities?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: Date;
}